refactor(sidekick): move sidekick command and internal package
Move `cmd/sidekick` from `internal/sidekick/cmd/sidekick` to
`cmd/sidekick`.

Flatten `internal/sidekick/internal/sidekick` into `internal/sidekick`,
since Go's `internal` packages cannot be imported from sibling
directories and all contents of that folder exist solely to support the
`sidekick` package.

Fixes https://github.com/googleapis/librarian/issues/1536
Fixes https://github.com/googleapis/librarian/issues/1548
