+++ cmd/sidekick/main_test.go
