+++ internal/sidekick/cmdline.go
