+++ internal/sidekick/cmdline_test.go
