+++ internal/sidekick/command.go
