+++ internal/sidekick/command_test.go
