+++ internal/sidekick/downloads_cache.go
