+++ internal/sidekick/downloads_cache_test.go
