+++ internal/sidekick/generate.go
