+++ internal/sidekick/refresh.go
