+++ internal/sidekick/refreshall.go
