+++ internal/sidekick/rust_generate.go
