+++ internal/sidekick/sidekick.go
