+++ internal/sidekick/sidekick_dart_test.go
