+++ internal/sidekick/sidekick_rust_prost_convert_test.go
