+++ internal/sidekick/sidekick_rust_prost_test.go
