+++ internal/sidekick/update.go
