+++ internal/sidekick/update_test.go
